requirements.txt
about-time==4.2.1
absl-py==2.2.0
accessible-pygments==0.0.5
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alabaster==1.0.0
alembic==1.15.1
alive-progress==3.2.0
annotated-types==0.7.0
anomalib==2.0.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.4
attrs==24.3.0
autograd==1.7.0
av==14.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
cma==3.2.2
colorama==0.4.6
comet_ml==3.49.5
comm==0.2.1
configobj==5.0.9
contourpy==1.3.1
coverage==7.7.1
cycler==0.11.0
databricks-sdk==0.47.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.21.2
dulwich==0.22.8
einops==0.8.1
everett==3.1.0
execnet==2.1.1
executing==0.8.3
fastapi==0.115.11
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.55.3
fqdn==1.5.1
FrEIA==0.2
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
gradio==5.22.0
gradio_client==1.8.0
grapheme==0.6.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
huggingface-hub==0.29.3
identify==2.6.9
idna==3.7
imageio==2.37.0
imagesize==1.4.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonargparse==4.37.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jstyleson==0.0.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.0
jupyter_server==2.14.1
jupyter_server_terminals==0.4.4
jupyterlab==4.3.6
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.2
linkify-it-py==2.0.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==2.0.4
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mlflow==2.21.0
mlflow-skinny==2.21.0
mpmath==1.3.0
multidict==6.2.0
myst-parser==4.0.1
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.16.4
nbformat==5.10.4
nbsphinx==0.9.7
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nncf==2.15.0
nodeenv==1.9.1
notebook==7.3.3
notebook_shim==0.2.3
numexpr==2.10.1
numpy==2.1.3
ollama==0.4.7
omegaconf==2.3.0
onnx==1.17.0
open-clip-torch==2.24.0
openai==1.68.2
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
openvino==2025.0.0
openvino-telemetry==2025.1.0
orjson==3.10.15
overrides==7.4.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.0
parso==0.8.4
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
pluggy==1.5.0
plumbum==1.9.0
ply==3.11
pre_commit==4.2.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.3.0
protobuf==5.29.4
psutil==5.9.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydata-sphinx-theme==0.15.4
pydot==2.0.0
pydub==0.25.1
Pygments==2.19.1
pymoo==0.6.1.3
pyparsing==3.2.0
pyproject-api==1.9.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-sugar==1.0.0
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2024.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.3
rsa==4.9
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==2.24.0
setproctitle==1.3.5
setuptools==72.1.0
shellingham==1.5.4
simplejson==3.20.1
sip==6.7.12
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==8.1.3
sphinx-autodoc-typehints==3.0.1
sphinx-book-theme==1.1.4
sphinx-copybutton==0.5.2
sphinx_design==0.6.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
stack-data==0.2.0
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.17.1
threadpoolctl==3.5.0
tifffile==2025.3.13
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchmetrics==1.7.0
torchvision==0.21.0+cu126
tornado==6.4.2
tox==4.24.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typeshed_client==2.7.0
typing_extensions==4.12.2
tzdata==2023.3
uc-micro-py==1.0.3
unicodedata2==15.1.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
waitress==3.0.2
wandb==0.19.8
wcwidth==0.2.5
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
wrapt==1.17.2
wurlitzer==3.1.1
yarl==1.18.3
